+++ lib/process_hub/initializer.ex
@@ -74,6 +74,7 @@ defmodule ProcessHub.Initializer do
         {Task.Supervisor, name: procs.task_sup},
         {ProcessHub.Coordinator, {hub_conf, procs, storage}},
         {ProcessHub.WorkerQueue, {hub_id, procs.worker_queue, storage.misc}},
+        {ProcessHub.PostStartWorker, {hub_id, procs.post_start_worker, storage.misc}},
         {ProcessHub.Janitor,
          {
            hub_id,
@@ -124,6 +125,7 @@ defmodule ProcessHub.Initializer do
       dist_sup: {:via, Registry, {system_registry, "dist_sup"}},
       task_sup: {:via, Registry, {system_registry, "task_sup"}},
       worker_queue: {:via, Registry, {system_registry, "worker_queue"}},
+      post_start_worker: {:via, Registry, {system_registry, "post_start_worker"}},
       janitor: {:via, Registry, {system_registry, "janitor"}}
     }
   end